#31 fetch services/rentals (test pass ✅)

+++ database/migrations/2024_04_07_070239_create_outers_table.php
@@ -22,7 +22,7 @@ return new class () extends Migration {
             $table = $this->groupOrgRelationship($table);
             $table->unsignedSmallInteger('shop_id')->nullable();
             $table->foreign('shop_id')->references('id')->on('shops');
-            $table->unsignedSmallInteger('product_id')->nullable();
+            $table->unsignedInteger('product_id')->nullable();
             $table->foreign('product_id')->references('id')->on('products');
             $table->string('slug')->unique()->collation('und_ns');
             $table->string('code')->index();
